feat: refactor imports to use util package for serverPlayers and gson
