+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/command/surfchat/SurfChatDeleteCommand.kt
@@ -4,10 +4,9 @@ import dev.jorel.commandapi.CommandAPICommand
 import dev.jorel.commandapi.kotlindsl.playerExecutor
 import dev.jorel.commandapi.kotlindsl.stringArgument
 import dev.slne.surf.chat.api.surfChatApi
-import dev.slne.surf.chat.bukkit.serverPlayers
+import dev.slne.surf.chat.bukkit.util.serverPlayers
 import dev.slne.surf.chat.core.service.historyService
 import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
-import org.bukkit.Bukkit
 import java.util.*
 
 class SurfChatDeleteCommand(commandName: String) : CommandAPICommand(commandName) {
