+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/service/BukkitDatabaseService.kt
@@ -2,29 +2,37 @@ package dev.slne.surf.chat.bukkit.service
 
 import com.github.benmanes.caffeine.cache.Caffeine
 import com.github.benmanes.caffeine.cache.LoadingCache
+
 import com.google.auto.service.AutoService
+
 import com.sksamuel.aedile.core.asLoadingCache
 import com.sksamuel.aedile.core.expireAfterWrite
 import com.sksamuel.aedile.core.withRemovalListener
+
 import dev.slne.surf.chat.api.model.BlacklistWordModel
 import dev.slne.surf.chat.api.model.ChatUserModel
 import dev.slne.surf.chat.api.model.HistoryEntryModel
-import dev.slne.surf.chat.bukkit.gson
 import dev.slne.surf.chat.bukkit.model.BukkitBlacklistWord
 import dev.slne.surf.chat.bukkit.model.BukkitChatUser
 import dev.slne.surf.chat.bukkit.model.BukkitHistoryEntry
 import dev.slne.surf.chat.bukkit.plugin
+import dev.slne.surf.chat.bukkit.util.gson
 import dev.slne.surf.chat.core.service.DatabaseService
 import dev.slne.surf.database.DatabaseProvider
 import dev.slne.surf.surfapi.core.api.util.toObjectList
 import dev.slne.surf.surfapi.core.api.util.toObjectSet
+
 import it.unimi.dsi.fastutil.objects.ObjectArraySet
 import it.unimi.dsi.fastutil.objects.ObjectList
 import it.unimi.dsi.fastutil.objects.ObjectSet
+
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+
 import net.kyori.adventure.util.Services.Fallback
+
 import org.bukkit.Bukkit
+
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
@@ -32,7 +40,9 @@ import org.jetbrains.exposed.sql.SqlExpressionBuilder.greaterEq
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.like
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
 import org.jetbrains.exposed.sql.transactions.transaction
+
 import java.util.*
+
 import kotlin.time.Duration.Companion.hours
 import kotlin.time.Duration.Companion.minutes
