+++ surf-chat-bukkit/src/main/kotlin/dev/slne/surf/chat/bukkit/util/CompletedComponents.kt
@@ -1,8 +1,6 @@
 package dev.slne.surf.chat.bukkit.util
 
 import dev.slne.surf.chat.api.model.ChannelModel
-import dev.slne.surf.chat.bukkit.plugin
-import dev.slne.surf.surfapi.core.api.messages.adventure.appendNewline
 import dev.slne.surf.surfapi.core.api.messages.adventure.buildText
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.event.ClickEvent
